% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amplifDat}
\alias{amplifDat}
\title{Gene amplification status}
\format{
\subsection{\code{amplifDat}}{

A matrix with 3469 rows and 39 columns.
}
}
\usage{
amplifDat
}
\description{
A numeric matrix of 0s and 1s of the same dimension as genomDat.
...
}
\keyword{datasets}
