% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno}
\alias{anno}
\title{Gene information}
\format{
\subsection{\code{anno}}{

A data.frame with 3469 rows and 2 columns.
}
}
\usage{
anno
}
\description{
A data.frame containing gene information.
...
}
\keyword{datasets}
