% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chrInfo}
\alias{chrInfo}
\title{Chromosomal information}
\format{
\subsection{\code{chrInfo}}{

A data.frame with 24 rows and 3 columns.
}
}
\usage{
chrInfo
}
\description{
A data.frame containing information for each chromosome.
...
}
\keyword{datasets}
