% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapRelated.R
\name{colorBar}
\alias{colorBar}
\title{Calibration bar for color images.}
\usage{
colorBar(
  x,
  horizontal = TRUE,
  col = grDevices::heat.colors(50),
  scale = 1:length(x),
  k = 10,
  cexAxis = 1,
  ...
)
}
\arguments{
\item{x}{.}

\item{horizontal}{.}

\item{col}{.}

\item{scale}{.}

\item{k}{.}

\item{cexAxis}{.}

\item{...}{additional arguments.}
}
\value{
A color vector.
}
\description{
This function produces a color image (color bar)
}
