% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapRelated.R
\name{cutCluster}
\alias{cutCluster}
\title{Cluster designation.}
\usage{
cutCluster(clustObj, ann, nClust = 2, rev = F)
}
\arguments{
\item{clustObj}{a hclust object.}

\item{ann}{a data.frame containing annotation for each observation.}

\item{nClust}{number of groups desired.}

\item{rev}{reverse the order of observations.}
}
\value{
a data.frame.
}
\description{
Cuts the tree into desired number of groups.
}
