% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_heatmap.R
\name{generate_heatmap}
\alias{generate_heatmap}
\title{Generate a Heat Map.}
\usage{
generate_heatmap(
  x,
  col_lab = c(FALSE, TRUE),
  row_lab = c(FALSE, TRUE),
  col_lab_vtr = NULL,
  row_lab_vtr = NULL,
  col_anno = c(FALSE, TRUE),
  row_anno = c(FALSE, TRUE),
  col_info = NULL,
  row_info = NULL,
  col_anno_var = NULL,
  row_anno_var = NULL,
  col_var_info = NULL,
  row_var_info = NULL,
  col_dend = c(FALSE, TRUE),
  row_dend = c(FALSE, TRUE),
  col_anno_name = NULL,
  row_anno_name = NULL,
  col_clust = NULL,
  row_clust = NULL,
  plot_info = list(margins = c(0.5, 0.5), cexCol = NULL, cexRow = NULL, cexColSide =
    NULL, cexRowSide = NULL, colorCatCol = NULL, colorCatRow = NULL, colorContCol = NULL,
    colorContRow = NULL),
  file_name = NULL,
  h_title = NULL,
  input_legend = c(FALSE, TRUE),
  legend_title = NULL,
  plotHeatmap = TRUE,
  heatmap_color = c("red", "blue", "grey"),
  zlm = c(-0.5, 0.5),
  ...
)
}
\arguments{
\item{x}{numeric matrix; values to be plotted in the heatmap.}

\item{col_lab}{boolean (T, F); to display column labels; default is FALSE.}

\item{row_lab}{boolean (T, F); to display row labels; default is FALSE.}

\item{col_lab_vtr}{character vector; with column labels to use; defaults to colnames(x).}

\item{row_lab_vtr}{character vector; with row labels to use; defaults to rownames(x).}

\item{col_anno}{boolean (T, F); to display column annotations; defaults to FALSE.}

\item{row_anno}{boolean (T, F); to display row annotations; defaults to FALSE.}

\item{col_info}{data frame; data to be used in column annotations; default is NULL.}

\item{row_info}{data frame; data to be used in row annotations; default is NULL.}

\item{col_anno_var}{character vector; with specific column annotation variables to see in column annotations; defaults to colnames(col_info).}

\item{row_anno_var}{character vector; with specific row annotation variables to see in row annotations; defaults to colnames(row_info).}

\item{col_var_info}{nested list; of column annotation variables that need specific colors and placement; options are "color", "level" for characters, "limit" for numerics and "mode"=c("character","numeric"); default is NULL.}

\item{row_var_info}{nested list; of row annotation variables that need specific colors and placement; options are "color", "level" for characters, "limit" for numerics and "mode"=c("character","numeric"); default is NULL.}

\item{col_dend}{boolean (T, F); to display column dendrogram; default is FALSE.}

\item{row_dend}{boolean (T, F); to display row dendrogram; default is FALSE.}

\item{col_anno_name}{character vector; to name column color bars; default is NULL.}

\item{row_anno_name}{character vector; to name row color bars; default is NULL.}

\item{col_clust}{variable of hclust or dendrogram class; to cluster by the column; default NULL when col_dend is FALSE.}

\item{row_clust}{variable of hclust or dendrogram class; to cluster by the row; default NULL when row_dend is FALSE.}

\item{plot_info}{list; of 8 variables to control heatmap and annotation column/row label size (in that order); default is list( margins=c(0.5,0.5), "cexCol" = NULL, "cexRow" = NULL, "cexColSide" = NULL, "cexRowSide" = NULL, colorCatCol=NULL, colorCatRow=NULL, colorContCol=NULL, colorContRow=NULL); 'cex': heatmap labels and 'cexSide': annotation labels.}

\item{file_name}{string; denotes the name of the file, must contain one of the following file types: .pdf, .jpeg, .png, .tiff; default is NULL.}

\item{h_title}{string; gives the heatmap output a title; default is NULL.}

\item{input_legend}{boolean (T, F); to display annotations legend; row, column or both will be added if selected; default is FALSE.}

\item{legend_title}{string; denotes the title of the legend; default is NULL.}

\item{plotHeatmap}{logical indicating if heatmap will be plotted; defaults to TRUE.}

\item{heatmap_color}{character vector of length 3; colors for high, low and middle values respectively; default is c("red","blue","grey").}

\item{zlm}{numeric vector of length 2; the minimum and maximum x values for which colors should be plotted; default is c(-0.5, 0.5).}

\item{...}{additional arguments.}
}
\value{
A list.
}
\description{
A heat map is a false color image (basically image(t(x))) with a dendrogram added to the left side and to the top. Typically, reordering of the rows and columns according to some set of values (row or column means) within the restrictions imposed by the dendrogram is carried out. This package takes the original heatmap function and reduces the argument complexity.
}
