% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genomDat}
\alias{genomDat}
\title{Gene expression data}
\format{
\subsection{\code{genomDat}}{

A matrix with 3469 rows and 39 columns.
}
}
\usage{
genomDat
}
\description{
A numeric matrix with gene expression data with rows representing genes and columns representing subjects.
...
}
\keyword{datasets}
