% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapRelated.R
\name{getCluster}
\alias{getCluster}
\title{Hierarchical cluster analysis based on various distance methods.}
\usage{
getCluster(dat, distMethod, linkMethod, absolute = F)
}
\arguments{
\item{dat}{a numeric matrix.}

\item{distMethod}{distance function.}

\item{linkMethod}{agglomeration method.}

\item{absolute}{a logical value. TRUE means absolute value of correlation is to be returned.}
}
\value{
an object of class "hclust".
}
\description{
This function computes distance matrix and then performs hierarchical clustering.
}
