% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapRelated.R
\name{getCosineDist}
\alias{getCosineDist}
\title{Cosine distance based distance matrix.}
\usage{
getCosineDist(x)
}
\arguments{
\item{x}{a numeric matrix.}
}
\value{
an object of class "dist".
}
\description{
Computes cosine distances and returns them as a distance matrix.
}
