% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapRelated.R
\name{getDist}
\alias{getDist}
\title{Returns a distance matrix.}
\usage{
getDist(dat, method = "pearson", absolute = FALSE)
}
\arguments{
\item{dat}{a numeric matrix passed to function cor.}

\item{method}{a character string indicating which correlation coefficient is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{absolute}{a logical value. TRUE means absolute value of correlation is to be returned.}
}
\value{
an object of class "dist".
}
\description{
This function computes correlation coefficients and returns them as a distance matrix.
}
