% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapRelated.R
\name{getKappaDist}
\alias{getKappaDist}
\title{Cohen's kappa coefficient based distance matrix.}
\usage{
getKappaDist(x, type = "Cohen", absolute = FALSE)
}
\arguments{
\item{x}{a numeric matrix passed to function cor.}

\item{type}{a character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{absolute}{a logical value. TRUE means absolute value of correlation is to be returned.}
}
\value{
an object of class "dist".
}
\description{
Computes Cohen's kappa coefficients and returns them as a distance matrix.
}
