% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap4.R
\name{heatmap4}
\alias{heatmap4}
\title{Heatmap internal function.}
\usage{
heatmap4(
  x,
  imp = TRUE,
  Rowv = NA,
  Colv = NULL,
  distfun = stats::dist,
  hclustfun = stats::hclust,
  add.expr,
  symm = FALSE,
  revC = identical(Colv, "Rowv"),
  scale = "none",
  na.rm = TRUE,
  margins = c(5, 5),
  ColSideColors,
  RowSideColors,
  cexRow = 0.2 + 1/log10(nr),
  cexCol = 0.2 + 1/log10(nc),
  fontRow = 1,
  fontCol = 1,
  labRow = NULL,
  labCol = NULL,
  lineRow = NULL,
  lineCol = NULL,
  lineColor = "black",
  totalR = nr,
  totalC = nc,
  ncr = NA,
  ncc = NA,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  verbose = getOption("verbose"),
  methodR = "ward.D",
  methodC = "ward.D",
  zlm = c(-0.5, 0.5),
  high = "green",
  low = "red",
  mid = "black",
  addAmp = NULL,
  colAmp = NULL,
  cexAmp = 0.25,
  addText = NULL,
  cexText = 1,
  lwdRect = graphics::par("lwd"),
  lwidHeatmap = 4,
  lheiHeatmap = 4,
  lwidRowSide = 0.2,
  lheiColSide = 0.2,
  cexRowSide = 1,
  cexColSide = 1,
  densColor = NULL,
  sideLabRow = 4,
  sideLabCol = 1,
  layoutRespect = T,
  sideColSide = c("left", "right"),
  sideRowSide = c("bottom", "top"),
  ...
)
}
\arguments{
\item{x}{numeric matrix of the values to be plotted.}

\item{imp}{logical impute missing values if TRUE.}

\item{Rowv}{determines if and how the row dendrogram should be computed and reordered. Either a dendrogram or a vector of values used to reorder the row dendrogram or NA to suppress any row dendrogram (and reordering) or by default, NULL, see ‘Details’ below.}

\item{Colv}{determines if and how the column dendrogram should be reordered. Has the same options as the Rowv argument above and additionally when x is a square matrix, Colv = "Rowv" means that columns should be treated identically to the rows (and so if there is to be no row dendrogram there will not be a column one either).}

\item{distfun}{function used to compute the distance (dissimilarity) between both rows and columns. Defaults to dist.}

\item{hclustfun}{function used to compute the hierarchical clustering when Rowv or Colv are not dendrograms. Defaults to hclust. Should take as argument a result of distfun and return an object to which as.dendrogram can be applied.}

\item{add.expr}{expression that will be evaluated after the call to image. Can be used to add components to the plot.}

\item{symm}{logical indicating if x should be treated symmetrically; can only be true when x is a square matrix.}

\item{revC}{logical indicating if the column order should be reversed for plotting, such that e.g., for the symmetric case, the symmetry axis is as usual.}

\item{scale}{character indicating if the values should be centered and scaled in either the row direction or the column direction, or none. The default is "row" if symm false, and "none" otherwise.}

\item{na.rm}{logical indicating whether NA's should be removed.}

\item{margins}{numeric vector of length 2 containing the margins (see graphics::par(mar = *)) for column and row names, respectively.}

\item{ColSideColors}{(optional) character vector of length ncol(x) containing the color names for a horizontal side bar that may be used to annotate the columns of x.}

\item{RowSideColors}{(optional) character vector of length nrow(x) containing the color names for a vertical side bar that may be used to annotate the rows of x.}

\item{cexRow, }{positive number, used as cex.axis in for the row axis labeling. The default currently only uses number of rows.}

\item{cexCol}{positive number, used as cex.axis in for the column axis labeling. The default currently only uses number of columns.}

\item{fontRow}{positive number, used as font.axis in for the row axis labeling. The default is 1.}

\item{fontCol}{positive number, used as font.axis in for the column axis labeling. The default is 1.}

\item{labRow}{character vector with row labels to use; defaults to rownames(x).}

\item{labCol}{character vector with column labels to use; these defaults to colnames(x).}

\item{lineRow}{(optional) positive number vector, used to draw horiontal lines. The default is NULL.}

\item{lineCol}{(optional) positive number vector, used to draw vertical lines. The default is NULL.}

\item{lineColor}{(optional) Color of lines drawn with lineRow and lineCol.}

\item{totalR}{.}

\item{totalC}{.}

\item{ncr}{.}

\item{ncc}{.}

\item{main}{main title; defaults to none.}

\item{xlab, }{x-axis title; defaults to none.}

\item{ylab}{y-axis title; defaults to none.}

\item{verbose}{logical indicating if information should be printed.}

\item{methodR}{Agglomeration method for clustering rows.}

\item{methodC}{Agglomeration method for clustering columns.}

\item{zlm}{the minimum and maximum x values for which colors should be plotted.}

\item{high}{color for high values.}

\item{low}{color for low values.}

\item{mid}{color for middle values.}

\item{addAmp}{(optional) matrix 0s and 1s of the same size as x to be added to the heatmap}

\item{colAmp}{(optional) color of addAmp}

\item{cexAmp}{(optional) size of addAmp}

\item{addText}{(optional) character matrix of the same size as x to be added to the heatmap}

\item{cexText}{(optional) color of addText}

\item{lwdRect}{.}

\item{lwidHeatmap}{.}

\item{lheiHeatmap}{.}

\item{lwidRowSide}{.}

\item{lheiColSide}{.}

\item{cexRowSide}{.}

\item{cexColSide}{.}

\item{densColor}{.}

\item{sideLabRow}{.}

\item{sideLabCol}{.}

\item{layoutRespect}{.}

\item{sideColSide}{.}

\item{sideRowSide}{.}

\item{...}{additional arguments passed on to image, e.g., col specifying the colors.}
}
\value{
A List.
}
\description{
Called by generate_heatmap to create heatmap.
}
