% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapRelated.R
\name{heatmapColorBar}
\alias{heatmapColorBar}
\title{Returns a color bar.}
\usage{
heatmapColorBar(
  limit,
  cols = c("green", "red", "black"),
  main = NULL,
  marginHMCBar = NULL,
  cexAxisHMCBar = 1,
  ...
)
}
\arguments{
\item{limit}{.}

\item{cols}{.}

\item{main}{.}

\item{marginHMCBar}{.}

\item{cexAxisHMCBar}{.}

\item{...}{additional arguments.}
}
\value{
A color vector.
}
\description{
Produces a color image (color bar) which can be used for the legend to another color image obtained from heatmap4.
}
