% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phen}
\alias{phen}
\title{Phenotypic (patient information) data}
\format{
\subsection{\code{phen}}{

A data.frame with 39 rows and 7 columns.
}
}
\usage{
phen
}
\description{
A data.frame containing subject information.
...
}
\keyword{datasets}
