% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap4.R
\name{rect.hclust.my}
\alias{rect.hclust.my}
\title{rect.hclust.my.}
\usage{
rect.hclust.my(
  tree,
  k = NULL,
  which = NULL,
  x = NULL,
  h = NULL,
  border = 2,
  cluster = NULL,
  horiz = FALSE
)
}
\arguments{
\item{tree}{an object of the type produced by hclust.}

\item{k}{Scalar. Cut the dendrogram such that exactly k clusters are produced.}

\item{which}{A vector selecting the clusters around which a rectangle should be drawn. which selects clusters by number (from left to right in the tree). Default is which = 1:k.}

\item{x}{A vector selecting the clusters around which a rectangle should be drawn. x selects clusters containing the respective horizontal coordinates.}

\item{h}{Scalar. Cut the dendrogram by cutting at height h.}

\item{border}{Vector with border colors for the rectangles.}

\item{cluster}{Optional vector with cluster memberships as returned by stats::cutree(hclust.obj, k = k), can be specified for efficiency if already computed.}

\item{horiz}{Horizontal or vertical.}
}
\value{
(Invisibly) returns a list where each element contains a vector of data points contained in the respective cluster.
}
\description{
Draws rectangles around the branches of a dendrogram highlighting the corresponding clusters. First the dendrogram is cut at a certain level, then a rectangle is drawn around selected branches.
}
