% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapRelated.R
\name{sampleColorLegend}
\alias{sampleColorLegend}
\title{Returns a legend.}
\usage{
sampleColorLegend(
  tls,
  col = NULL,
  lty = NULL,
  border = NULL,
  pch = NULL,
  lwd = NULL,
  legendTitle = NULL,
  cex = NULL,
  density = NULL,
  ...
)
}
\arguments{
\item{tls}{.}

\item{col}{.}

\item{lty}{.}

\item{border}{.}

\item{pch}{.}

\item{lwd}{.}

\item{legendTitle}{.}

\item{cex}{.}

\item{density}{.}

\item{...}{additional arguments.}
}
\description{
This function can be used to add legends to row and column color bars.
}
